TASK: Test `ProjectionTransactionTrait` when using external projections
CASE: $this->dbal->isTransactionActive() === false
